+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Changelog for galois-field
 
+## 0.1.1
+
+* Add `Bits` instances
+* Add `Arbitary` instances
+* Move Arbitrary instances of galois fields, extensions and polynomials rings to ``src`
+* Add tasty-discover
+
 ## 0.1.0
 
 * Initial release.